* 'dev' of https://github.com/drmeister/clasp:
  Fixed problems with write-string for #314
  For #314 Removed vector virtual T_sp& operator[]
  Added function to register GC roots
  Split literal write into bclasp and cclasp version
  Added a load-time function to add GC roots

+++ src/core/arrayDisplaced.cc
@@ -91,7 +91,7 @@ void ArrayDisplaced_O::rowMajorAset(cl_index idx, T_sp value) {
   this->_Array->rowMajorAset(this->_DisplacedIndexOffset + idx, value);
 }
 
-T_sp ArrayDisplaced_O::aset_unsafe(int idx, T_sp value) {
+T_sp ArrayDisplaced_O::aset_unsafe(size_t idx, T_sp value) {
   this->_Array->aset_unsafe(this->_DisplacedIndexOffset + idx, value);
   return value;
 }
